+++ src/main/java/com/wipro/fhir/data/resource_model/FamilyMemberHistoryDataModel.java
@@ -42,9 +42,9 @@ public class FamilyMemberHistoryDataModel implements Serializable {
 
 	// @Id
 	// @GeneratedValue
-	private Long id;
-	private Long beneficiaryRegID;
-	private Long visitCode;
+	private BigInteger id;
+	private BigInteger beneficiaryRegID;
+	private BigInteger visitCode;
 	private Integer providerServiceMapID;
 	private Integer vanID;
 	private String familyMembers;
@@ -57,9 +57,9 @@ public class FamilyMemberHistoryDataModel implements Serializable {
 	}
 
 	public FamilyMemberHistoryDataModel(Object[] objArr) {
-		this.id = ((BigInteger) objArr[0]).longValue();
-		this.beneficiaryRegID = ((BigInteger) objArr[1]).longValue();
-		this.visitCode = ((BigInteger) objArr[2]).longValue();
+		this.id = BigInteger.valueOf((long) objArr[0]);
+		this.beneficiaryRegID = BigInteger.valueOf((long) objArr[1]);
+		this.visitCode = BigInteger.valueOf((long) objArr[2]);
 		this.providerServiceMapID = (Integer) objArr[3];
 		this.vanID = (Integer) objArr[4];
 		this.familyMembers = (String) objArr[5];
